astronomer/requirements.txt
astronomer-cosmos
apache-airflow-providers-trino